Refactor: change content component folder name
