+++ src/components/wrappers/ArtistWrapper.tsx
